[kernels] fixes Convolution kernel result handling
This patch adds the `isReturnValue` flag to our wrapper creation script
and to the 2nd and 3rd result value of the ConvolutionOp. There
was no way to have these values be passed as pointer, and also be
declared as an output value.

Additionally, adds a small check to verify the passed arguments of the
convolution kernel.