[kernels] fixes in-place calculation of `BiasAdd`
This patch changes the `BiasAdd` kernel to provide its result to the
dedicated output parameter. Before, the result was just aliasing the
input causing a segfault because of an use after free access. The
automatic memory management could deallocate the input matrix to the
`BiasAdd` before the result would be accessed.