Added the reading of Characteristics to the MetaData Parser.
For this we need to change the version of the json reader we are using, to enable the use of optional parameters. The optional Parameters i picked should be revisted AND over all checked if this io as is.
Added the Fields to the docu too. Here again a check should be done.