Minor change for clarity.
Let’s say there are two kinds of kernels: non-instrumentable ones (`createDaphneContext` and `destroyDaphneContext`) and instrumentable ones (all others). A kernel should get the `kId` parameter and do the pre/post kernel instrumentation if and only if it is instrumentable. With this simple rule, we can omit the special case for `destroyDaphneContext` in `genKernelInst.py` (which is good because this central script is already quite complex). It suffices to take `destroyDaphneContext` into account in `isInstrumentableOp()`. As a consequence, the generated _destroyDaphneContext`-kernel does not have the `kId` parameter anymore. To make it work end-to-end, we need to omit this parameter when creating the `CallKernelOp` in `RewriteToCallKernelOpPass.cpp` for all non-instrumentable operations. Thus, I slightly changed this file, too. Finally, the `DaphneContext` is still added as the last parameter for all kernels but `CreateDaphneContextOp`.