Fix parfor return lowering
The former version would first replace the operand of return with the output pointer and then replace the return by storing this new operand to the output pointer. By just not doing the first part we get a sensible translation where the original return operand is now written into the output pointer instead.