Re-Enable garbage collection and Fix parfor case
When the input/output matrix object is created it has only one use as the arg to parfor and therefore is decremented right after. The return value of parfor points to the same object though so we need to add an increment to make sure the input/output matrix is available after the execution.